+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+### Unreleased
+- `gpt-3.5-turbo-0613` and `gpt-3.5-turbo-16k-0613` is deprecated on 2024-09-13.  
+   You can still call these models, but it is removed from model name completion.
+
 ### 4.3.0
 - Add `-Include` param for `Get-ThreadRunStep`.
 - Add `-RankerForFileSearch` and `-ScoreThresholdForFileSearch` params for `New-Assistant`.